backend/pyproject.toml
[project]
name = "alphamesh-backend"
version = "0.1.0"
description = "AlphaMesh MVP backend scaffold"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.4.0",
    "psycopg[binary]>=3.2.0",
    "langchain-core>=1.3.2",
    "langchain-openai>=1.2.1",
    "langchain-anthropic>=1.4.1",
    "langchain-google-genai>=4.2.2",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
